🏗️ structure: changed directory structure of Box component
