+++ packages/box/src/index.ts
@@ -1,1 +1,1 @@
-export * from './box';
+export * from './components/Box';
